Revert "Merge branch 'dev' into fix/adapters"
This reverts commit 0151749fb9c67ff4712d55d757a2da0d984fe3ad, reversing
changes made to 7e731e5ae3510c5d95e4a193f8b78f0bfb634075.

+++ src/main/java/com/adondeband/back_end_adonde_band/dominio/jornada/JornadaPort.java
@@ -5,5 +5,5 @@ import java.util.List;
 public interface JornadaPort {
     Jornada save(Jornada jornada);
 
-    List<Jornada> findById(Long s);
+    List<Jornada> findById(int s);
 }
